Modified trainer classes to include a "trainBoth" method, allowing two networks to play against each other.
Fixed a potential issue where tdNeuralNetPlayer was being trained on
inverted valeus for end states.

Minor changes to tdneuralnetplayer to improve efficiency.

Added a Linear concrete class for Activation.

Added the ability to use a horizontal stretch in Sigmoid.

"Fixed" an issue in tdneuralnetplayer for connectFour where final
outputs converged to 1 or -1.  The issue still may occur; it was
somewhat rare before.

Updated ConnectFourMainWidget and TicTacToeMainWidget to include a
"train both" option.

Fixed many memory leaks and bugs with file loading and saving.